fix: Fix stopping of previously installed artifact
This commit addresses a bug that was causing installation failures when
installing a new artifact. The issue stemmed from attempting to stop the
previously installed artifact by checking for the existence of its "manifests"
directory. However, if the previous installation failed before creating the
"manifests" directory, the `STOP` operation would fail. This commit modifies
the condition for executing the `STOP` operation to now check for the existence
of the "manifests" directory itself, rather than its parent directory. This
adjustment ensures that the `STOP` operation is only performed when the
"manifests" directory actually exists.

Ticket: None
Changelog: Title
Signed-off-by: Paul Kretschel <slither-fable0w@icloud.com>